process adsorbates

+++ scripts/process_adsorbates.py
@@ -114,8 +114,10 @@ def process_adsorbates(root_dir, N_ADS=82):
             _ = uncompress_data(compressed_dir)
             shutil.rmtree(compressed_dir)
 
-        write_ads_to_lmdb(root_dir, i)
-        shutil.rmtree(ads_dir)
+        if not os.path.exists(os.path.join(root_dir, str(i)+".lmdb")):
+            write_ads_to_lmdb(root_dir, i)
+        if os.path.exists(ads_dir):
+            shutil.rmtree(ads_dir)
         logging.info("Finished adsorbate %d" % i)
